Fix multiple issues with execution
A requirements-file was added.
Furthermore a gracefull shutdown handler.
Additionally the strategy for the DevDataServer was adjusted.
